[chore] : Moya 코드 전체 주석처리, InfoView, ViewModel 임시병합
